requirements.txt
boto3
chalice
PyJWT
pydantic
